Refactor single config entry warning: move function to dedicated file and update imports
